src/requirements.txt
boto3==1.18.44
botocore==1.21.44
certifi==2021.5.30
charset-normalizer==2.0.6
future==0.18.2
idna==3.2
jmespath==0.10.0
oauthlib==3.1.1
python-dateutil==2.8.2
python-json-logger==2.0.2
python-twitter==3.5
PyYAML==5.4.1
requests==2.26.0
requests-oauthlib==1.3.0
s3transfer==0.5.0
six==1.16.0
urllib3==1.26.6
